+++ src/AstroPhotography/api/split.py
@@ -5,7 +5,7 @@ from ..core.logger import logger
 from ..core.RawConv import RawConv
 from ..core.file_writer import file_writer
 
-def main(rawfile, output, black, extension):
+def main(rawfile, output, keepblack, extension):
     """ Execute the split command.
     
     :param rawfile: RAW input file to process.
@@ -13,24 +13,25 @@ def main(rawfile, output, black, extension):
       file names will consist of the prefix followed by '_r.', 
       '_g1.', '_b.' and '_g2.', followed by the file name
       extension.
-    :param black: If true the camera black levels will be subtracted
-      from the channel data.
+    :param keepblack: If true the camera black levels (similar to a CCD bias?)
     :param extension: File type and extension to use, e.g. 'png', 'jpg'
       'jp2', or 'fits'
     """
     logger.info("Executing split command on {:s}".format(rawfile))
     
+    subblack = not keepblack
+    
     # Read and process the file.
     rawconv = RawConv(rawfile)
-    r_im, g1_im, b_im, g2_im = rawconv.split(subtract_black=black)
+    r_im, g1_im, b_im, g2_im, exif_dict = rawconv.split(subtract_black=subblack)
     
     r_fname  = output + '_r.'  + extension
     g1_fname = output + '_g1.' + extension
     b_fname  = output + '_b.'  + extension
     g2_fname = output + '_g2.' + extension
     
-    file_writer(r_im,  r_fname)
-    file_writer(g1_im, g1_fname)
-    file_writer(b_im,  b_fname)
-    file_writer(g2_im, g2_fname)
+    file_writer(r_fname,  r_im,  exif_dict)
+    file_writer(g1_fname, g1_im, exif_dict)
+    file_writer(b_fname,  b_im,  exif_dict)
+    file_writer(g2_fname, g2_im, exif_dict)
     return